+++ src/app/(providers)/(root)/(goodsShop)/shop/order/page.tsx
@@ -1,11 +1,10 @@
 import OrderPageSection from '@/components/order/OrderPageSection';
-import Page from '@/components/pages/Page';
 
 const OrderPage = () => {
   return (
-    <Page>
+    <>
       <OrderPageSection />
-    </Page>
+    </>
   );
 };
