+++ src/app/(providers)/(root)/(goodsShop)/shop/page.tsx
@@ -1,15 +1,15 @@
-import Page from '@/components/pages/Page';
 import GoodsList from '@/components/shop/GoodsList';
+import { orbitron } from '../../../../../../public/fonts/orbitron';
 
 const ShopPage = () => {
   return (
     <>
-      <Page>
-        <h1 className='border-b-2 border-white text-4xl font-bold text-white mt-20 py-4'>
-          GOODS SHOP
-        </h1>
-        <GoodsList />
-      </Page>
+      <h1
+        className={`border-b-2 border-white text-4xl font-bold text-white mt-20 py-4 ${orbitron.className}`}
+      >
+        GOODS SHOP
+      </h1>
+      <GoodsList />
     </>
   );
 };
