+++ src/app/(providers)/(root)/(goodsShop)/shop/payment/[orderId]/page.tsx
@@ -1,14 +1,16 @@
 import PaymentWidget from '@/components/order/payment/PaymentWidget';
-import Page from '@/components/pages/Page';
+import { orbitron } from '../../../../../../../../public/fonts/orbitron';
 
 const PaymentPage = () => {
   return (
-    <Page>
-      <h1 className='border-b-[1px] border-white text-xl font-bold text-white mt-20 py-4'>
+    <>
+      <h1
+        className={`border-b-[1px] border-white text-xl font-bold text-white mt-20 py-4 ${orbitron.className}`}
+      >
         결제하기
       </h1>
       <PaymentWidget />
-    </Page>
+    </>
   );
 };
