+++ src/app/(providers)/(root)/(goodsShop)/shop/payment/success/page.tsx
@@ -1,11 +1,10 @@
 import SuccessPaymentSection from '@/components/order/payment/success/SuccessPaymentSection';
-import Page from '@/components/pages/Page';
 
 const SuccessPage = () => {
   return (
-    <Page>
+    <>
       <SuccessPaymentSection />
-    </Page>
+    </>
   );
 };
