+++ src/app/(providers)/(root)/(mypage)/wishlist/page.tsx
@@ -3,9 +3,7 @@ import TabSelector from '@/components/mypage/wishlist/TabSelector';
 const WishListPage = () => {
   return (
     <>
-      <div>
-        <TabSelector />
-      </div>
+      <TabSelector />
     </>
   );
 };
