+++ src/components/shop/detail/GoodsInfo.tsx
@@ -4,8 +4,8 @@ import ShareIcon32px from '@/components/common/icons/32px/ShareIcon32px';
 import Image from 'next/image';
 import QuantityBtn from './QuantityBtn';
 import useAuthStore from '@/zustand/store/useAuth';
-import Hearts from '../Hearts';
 import { Tables } from '@/types/supabase';
+import GoodsHearts from '../GoodsHearts';
 
 type GoodsInfoProps = {
   goods: Tables<'goods'>;
@@ -70,7 +70,7 @@ const GoodsInfo = ({ goods, goods_id }: GoodsInfoProps) => {
         </div>
         <div className='gap-4 flex mt-5 w-full'>
           <div className='flex p-2 rounded-lg items-center border-2 border-solid border-primary-400'>
-            <Hearts goods_id={goods_id} user_id={user?.id} />
+            <GoodsHearts goods_id={goods_id} user_id={user?.id} />
           </div>
           <div className='flex flex-grow gap-4 text-base h-[53px]'>
             <button className='border-solid border-2 w-full border-primary-400 rounded-lg'>
