+++ src/components/tour/tourDetail/TourHearts.tsx
@@ -3,15 +3,12 @@
 import Loading from '@/components/common/Loading';
 import HeartDefaultIcon32px from '@/components/common/icons/32px/HeartDefaultIcon32px';
 import HeartPressedIcon32px from '@/components/common/icons/32px/HeartPressedIcon32px';
-import { useGetLikedGoodsByUser, useToggleLikeGoods } from '@/hooks/goodsHooks';
 import {
   useGetIsLikedTourByUser,
   useToggleLikeTours,
-} from '@/hooks/toursHooks';
+} from '@/hooks/apis/tours.api';
 import { toggleLikeToursParamsType } from '@/types/tour';
 import toast from 'react-hot-toast';
-import { IoHeart } from 'react-icons/io5';
-import { IoHeartOutline } from 'react-icons/io5';
 
 interface TourHeartsProps {
   tour_id: string;
@@ -25,11 +22,7 @@ function TourHearts({ tour_id, user_id }: TourHeartsProps) {
     isPending,
   } = useGetIsLikedTourByUser(tour_id, user_id!);
 
-  const { mutate: likeMutate } = useToggleLikeTours(
-    tour_id,
-    user_id!,
-    isLiked!,
-  );
+  const { mutate: likeMutate } = useToggleLikeTours(tour_id, user_id!);
 
   const handleToggleLike = () => {
     if (!user_id) return toast.error('로그인 해주세요!');
